scco/ml_generation/ml_models/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
numpy==1.26.4
files==1.1.1